refactor: move models directory to top level
