+++ src/daemons/Synchronizer.ts
@@ -46,19 +46,19 @@ import {
     UserTransitionProof,
 } from '@unirep/contracts'
 import { EventEmitter } from 'events'
-import Proof from '../database/models/proof'
-import UserSignUp from '../database/models/userSignUp'
-import GSTLeaf from '../database/models/GSTLeaf'
-import GSTRoot from '../database/models/GSTRoots'
-import Nullifier from '../database/models/nullifiers'
-import Attestation from '../database/models/attestation'
-import Epoch from '../database/models/epoch'
-import Record from '../database/models/record'
-import EpkRecord from '../database/models/epkRecord'
-import Post from '../database/models/post'
-import Comment from '../database/models/comment'
-import BlockNumber from '../database/models/blockNumber'
-import SynchronizerState from '../database/models/synchronizerState'
+import Proof from '../models/proof'
+import UserSignUp from '../models/userSignUp'
+import GSTLeaf from '../models/GSTLeaf'
+import GSTRoot from '../models/GSTRoots'
+import Nullifier from '../models/nullifiers'
+import Attestation from '../models/attestation'
+import Epoch from '../models/epoch'
+import Record from '../models/record'
+import EpkRecord from '../models/epkRecord'
+import Post from '../models/post'
+import Comment from '../models/comment'
+import BlockNumber from '../models/blockNumber'
+import SynchronizerState from '../models/synchronizerState'
 
 const encodeBigIntArray = (arr: BigInt[]): string => {
     return JSON.stringify(stringifyBigInts(arr))
