+++ src/models/epkRecord.ts
