+++ src/models/proof.ts
