+++ src/routers/RecordRouter.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response, Router } from 'express'
-import EpkRecord from '../database/models/epkRecord'
-import Record from '../database/models/record'
-import userSignUp from '../database/models/userSignUp'
+import EpkRecord from '../models/epkRecord'
+import Record from '../models/record'
+import userSignUp from '../models/userSignUp'
 import RecordController from '../controllers/RecordController'
 import { ActionType } from '../constants'
 import catchError from './catchError'
